+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ### Added
 - Header comment check to pr checklist
 - Explicit typing ESLint rule for useState
+- ESLint rule for TODO and FIXME comments
+
 ### Changed
 
 ### Fixed